Add files via upload
This is a simple example. We can develop and try on it. In order for it to work correctly on your device, open the cmd, navigate to the folder which contains the files and run these commands: Components

yarn add @chatscope/chat-ui-kit-react
or
npm install @chatscope/chat-ui-kit-react
Styles

yarn add @chatscope/chat-ui-kit-styles;
or
npm install @chatscope/chat-ui-kit-styles;